% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devl_connect.R
\name{devl_connect}
\alias{devl_connect}
\title{Connection function to DEVL database}
\usage{
devl_connect()
}
\value{
A String connection to Microsoft SQL Server
}
\description{
This function connects to the devl database according to DEVL schema
}
\examples{
devl_connect()
}
