% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreaconnect-package.R
\docType{package}
\name{dreaconnect-package}
\alias{dreaconnect}
\alias{dreaconnect-package}
\title{dreaconnect: Connection tools to department databases}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stevens Dormezil \email{steves.dormezil@palmbeachschools.org} (\href{https://orcid.org/0000-0003-1886-6964}{ORCID})

}
\keyword{internal}
