% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edw_connect.R
\name{edw_connect}
\alias{edw_connect}
\title{EDW Connection Function}
\usage{
edw_connect()
}
\value{
A string connection to Microsoft SQL Server
}
\description{
This function connects to the edw database according to edw schema
}
\examples{
edw_connect()
}
