% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edw_dictionary.R
\name{edw_dictionary}
\alias{edw_dictionary}
\title{Title List All Variables in Table Dictionary}
\usage{
edw_dictionary()
}
\value{
Tibble of edw dictionary list
}
\description{
This function list of all edw tables and their variables
}
\section{Last Updated By}{

Stevens Dormezil
}

\section{Last Update Date}{

2022/06/08
}

\examples{
edw_dictionary()

}
