% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\title{List EDW Tables}
\usage{
list_tables(database = "edw")
}
\arguments{
\item{database}{SQL database connection string}
}
\value{
Tibble of table list
}
\description{
This function list edw and devl tables. Defaults to edw.
}
\section{Last Updated By}{

Stevens Dormezil
}

\section{Last Update Date}{

2022/06/08
}

\examples{
list_tables()
list_tables("devl")

}
