% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_connect.R
\name{table_connect}
\alias{table_connect}
\title{Function to connect to table}
\usage{
table_connect(con, table_name = "stu_indic_fact")
}
\arguments{
\item{table_name}{}
}
\value{
A string connection to the table.
}
\description{
This function helps connect to tables
}
\examples{
location_dim_db <- table_connect("location_dim")
}
