% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreaconnect-package.R
\docType{package}
\name{dreaconnect-package}
\alias{dreaconnect}
\alias{dreaconnect-package}
\title{dreaconnect: Connection tools to department databases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Dreaconnect helps ease database connections base on drea database schema.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dormezil/dreaconnect}
  \item Report bugs at \url{https://github.com/dormezil/dreaconnect/issues}
}

}
\author{
\strong{Maintainer}: Stevens Dormezil \email{steves.dormezil@palmbeachschools.org} (\href{https://orcid.org/0000-0003-1886-6964}{ORCID})

}
\keyword{internal}
