% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_connect.R
\name{table_connect}
\alias{table_connect}
\title{Function to connect to table}
\usage{
table_connect(con, table_name = "stu_indic_fact", database = "edw")
}
\arguments{
\item{con}{String to previously made odbc connection}

\item{table_name}{String of table to connect to}

\item{database}{Database to connect to. Options include "edw" and "devl"}
}
\value{
A string connection to the table.
}
\description{
This function helps connect to tables
}
\section{Last Updated By}{

Stevens Dormezil
}

\section{Last Update Date}{

2022/06/08
}

\examples{
\dontrun{location_dim_db <- table_connect("location_dim")}

}
